Add files via upload
Boucle pour le choix de mode et modification de float à int de la fréquence

+++ V_Interface.h
@@ -7,7 +7,7 @@ class Interface {
 public:
     Interface();
     std::string ChooseMode();
-    float ChooseUpdateInterval();
+    int ChooseUpdateInterval();
     bool ChooseToroidal();
     std::string get_display_mode() const;
     int get_update_interval() const;
